This is how we invoke the plugin
The function has been stored in $.fn now, instead of our Globals object. Similarly, near the end, we store our default color in $.fn.stripe.

The first parameter has been removed, since that will now be the query upon which our plugin acts. Because of that, within the code, the $(query) has been replaced with this.

Before we start finding tables, we override the defaults with any incoming options.

To properly behave as a jQuery collection method, we return this at the end.